Add IRC bot remote control service.
This is a generic `brpc` service that, in the future, can be used to
allow external workers to provide services over the bot, hopefully
allowing a sort of a hot reload capability. For now, it simply serves
a way to gain low-level control over the bot's IRC connection,
allowing watching messages and sending new ones, in case the bot needs
to (e.g.) join/part or talk to services.