Add nick/channel tracking to IRC bot framework.
For correct per-channel logging of NICK and QUIT messages (which are
sent just once, even if you have multiple channels in common, and do
not mention channel name), we need to start tracking the
nickname/channel association. This commit adds rudimentary support for
that in the framework.

Regarding processing order, JOIN and PART messages (and the initial
list of names sent upon joining) are processed before passing messages
down to any modules, but NICK and QUIT messages are handled after the
fact, so that the channel associations of the sender of the message
are still present.  Arguably the same *could* be done for JOIN and
PART as well, by redefining the semantics to make those events mean
that the action in question happens right after the message.